requirements.txt
imutils
picamera
simplejson
cv2
tornado
pycrypto